refactor: modularize temporal-bun bridge
Refs #1477
